chore: added default layout to all pages
